+++ instrumentation/net/http/otelhttp/internal/request/resp_writer_wrapper_test.go
@@ -1,7 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package internal
+package request
 
 import (
 	"net/http"
